Add files via upload
edX Microsoft DEV330x - Introduction to Python: Creating Scalable, Robust, Interactive Code

This course was archived and not available for a certificate as of April 4, 2020 by the time I got to it on edX.  It is the next course in the sequence following DEV236x Introduction to Python: Absolute Beginner and DEV274x Introduction to Python: Fundamentals.

I completed auditing the course on May 2, 2020 including the practice notebooks, quizzes, midterms and final.  I decided to upload the Jupyter notebooks as proof of completion and to practice using GitHub.  Furthermore, attached at the bottom of this repository is a screenshot PDF showing completion of course requirements.

I thoroughly enjoyed the 3 Python courses referenced above (created and taught by Microsoft) and believe they gave me a solid intro to programming in Python.